preencher criador do evento no service

+++ app/src/main/java/com/ufrpe/bsi/soresenha/usuario/negocio/UsuarioServices.java
@@ -34,7 +34,7 @@ public class UsuarioServices {
         usuarioDAO.update(usuario);
     }
 
-    public Usuario getByID(int id) {
+    public Usuario getByID(long id) {
         return usuarioDAO.getByID(id);
     }
 }
